Basic summary statistics and input table
Now the input is in the form of a table with no header, where the first
column is the id of the run and the second column is the path to the
file. The program outputs a tsv summary table with the statistics from
each input file. Currently the statistics are:
    - Number of repeats
    - Number of unique repeats